pyproject.toml
[tool.poetry]
name = "tg_bot"
version = "0.1.0"
description = ""
authors = ["konstantin <arkov.constantin@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.0a4"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
nats-py = "^2.1.7"
install = "^1.3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
